Simplify

+++ src/deployer/SafeGuard.sol
@@ -386,22 +386,21 @@ abstract contract ZeroExSettlerDeployerSafeGuardBase is IGuard {
             // Forbid calls to `this.checkTransaction` and `this.checkAfterExecution`.
             if (
                 data.length >= 68
-                    && ((data.length >= 356
-                            && uint256(uint32(bytes4(data))) == uint256(uint32(this.checkTransaction.selector)))
-                        || uint256(uint32(bytes4(data))) == uint256(uint32(this.checkAfterExecution.selector)))
+                    && ((data.length >= 356 && uint32(bytes4(data)) == uint32(this.checkTransaction.selector))
+                        || uint32(bytes4(data)) == uint32(this.checkAfterExecution.selector))
             ) {
                 revert ForbiddenCall(callsCount, to, data);
             }
             // Transactions containing calls to `this.unlock()` must be unanimous
-            requireUnanimity = requireUnanimity
-                || (data.length >= 4 && uint256(uint32(bytes4(data))) == uint256(uint32(this.unlock.selector)));
+            requireUnanimity =
+                requireUnanimity || (data.length >= 4 && uint32(bytes4(data)) == uint32(this.unlock.selector));
         }
         return requireUnanimity;
     }
 
     /// See comment in `checkTransaction`
     function _checkDelegateCall(bool requireUnanimity, address to, bytes calldata data) private view returns (bool) {
-        if (to == _MULTISEND && uint256(uint32(bytes4(data))) == uint256(uint32(ISafeMultiSend.multiSend.selector))) {
+        if (to == _MULTISEND && uint32(bytes4(data)) == uint32(ISafeMultiSend.multiSend.selector)) {
             // Slice off the selector.
             bytes calldata multicalls = data[4:];
             // Follow the dynamic-type ABIencoding indirection to the `transactions` argument.
@@ -437,7 +436,7 @@ abstract contract ZeroExSettlerDeployerSafeGuardBase is IGuard {
                     // invariants in between each call.
                     if (
                         multicallTo != address(this) || multicallData.length < 4
-                            || uint256(uint32(bytes4(multicallData))) != uint256(uint32(this.check.selector))
+                            || uint32(bytes4(multicallData)) != uint32(this.check.selector)
                     ) {
                         revert GuardCheckNotEnforced(callsCount, multicallTo, multicallData);
                     }
